refactor(topic): change path import
